# backend_express/swagger.yaml
openapi: 3.0.3
info:
  title: VPP Aggregation Platform API
  description: |
    API documentation for the Virtual Power Plant (VPP) Aggregation Platform.
    This platform manages edge nodes, telemetry data, dispatch commands, market trading, and multi-agent coordination.
  version: 1.0.0
  contact:
    name: API Support
    email: amansharma12607@gmail.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.vpp-platform.com/api
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login or /api/vendor-auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    DispatchRequest:
      type: object
      required:
        - targets
        - action
      properties:
        targets:
          oneOf:
            - type: string
              enum: [all]
            - type: array
              items:
                type: string
            - type: string
        action:
          type: string
          enum: [charge, discharge, defer_load, hold]
        params:
          type: object
    DispatchResponse:
      type: object
      properties:
        success:
          type: boolean
        dispatched:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        timestamp:
          type: string
          format: date-time
    DispatchLog:
      type: object
      properties:
        dc_id:
          type: string
        action:
          type: string
        params:
          type: object
        issued_by:
          type: string
        operator_email:
          type: string
        timestamp:
          type: string
          format: date-time
        status:
          type: string
        result:
          type: object
...
    DispatchPlan:
      type: object
      properties:
        plan:
          type: object
        expected_revenue:
          type: number
        confidence:
          type: number
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /dispatch:
    post:
      tags:
        - Dispatch
      summary: Dispatch command to edge nodes
      description: Send dispatch commands to one or more edge nodes
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRequest'
      responses:
        '200':
          description: Command dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dispatch/logs:
    get:
      tags:
        - Dispatch
      summary: Get dispatch history
      description: Retrieve dispatch command history
      security: []
      parameters:
        - name: dc_id
          in: query
          schema:
            type: string
          description: Filter by data center ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of records
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
      responses:
        '200':
          description: Dispatch logs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DispatchLog'
...
  /optimization/dispatch:
    post:
      tags:
        - Optimization
      summary: Optimize and execute dispatch
      description: Generate optimized dispatch plan and optionally execute it
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                auto_execute:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Dispatch optimized
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DispatchPlan'
...